Add files via upload
About Intervention, you can got to it's page to copy and see the tutorial, simply write composer require intervention/images i guess (the php composer.phar didn't work for me), then go to config/app.php, there you need to add 2 lines, they're on the site.
On the comments, I've added the filter for replies whose parents are blocked, and also count the comments on the pRequest
Code: on request-description there's code on the comments count, and the if for replies (hasBlockedParent verification). The Comment has that method.
On the edit of profiles, I've done what I could, which was whenever you click "edit" on the profile it takes you to the edit-create-profile.blade (2 routes on web.php). There you get to see your already filled info (shit forgot the presentation), and yo can click to upload an image (I think it works, the dd gave true on hasFile). Anyways, UserController has 2 methods, profilepage has the edit option and uncommented show image, they're above the isAdmin check, the edi